// lib/types/reports/types.ts
export type ReportView = 'machines' | 'locations' | 'meters';

export type ReportTab = {
  id: ReportView;
  label: string;
  icon?: string;
  description?: string;
};